refined implementation and addressed feedback
