+++ exporthistory/src/main/java/com/microsoft/durabletask/exporthistory/ExportJob.java
@@ -19,7 +19,7 @@ import java.util.stream.Collectors;
  * Operations are dispatched by method name (case-insensitive): {@code Create}, {@code Get}, {@code Run},
  * {@code CommitCheckpoint}, {@code MarkAsCompleted}, {@code MarkAsFailed}, {@code Delete}.
  */
-public final class ExportJob extends AbstractTaskEntity<ExportJobState> {
+final class ExportJob extends AbstractTaskEntity<ExportJobState> {
 
     /** The registered entity name. */
     public static final String NAME = "ExportJob";
@@ -64,10 +64,6 @@ public final class ExportJob extends AbstractTaskEntity<ExportJobState> {
                         ? null
                         : new ArrayList<>(creationOptions.getRuntimeStatus());
 
-        // Resolve the completion-time lower bound to the job creation instant when the caller omitted it. This
-        // mirrors the .NET SDK (completedTimeFrom ?? UtcNow) so a CONTINUOUS job tails only completions that occur
-        // after the job is created, instead of re-exporting every historical instance in the task hub. BATCH always
-        // supplies an explicit lower bound (enforced by ExportJobCreationOptions.validateForCreate()).
         Instant now = Instant.now();
         Instant completedTimeFrom = creationOptions.getCompletedTimeFrom() != null
                 ? creationOptions.getCompletedTimeFrom()
