+++ exporthistory/src/main/java/com/microsoft/durabletask/exporthistory/ListTerminalInstancesActivity.java
@@ -5,16 +5,20 @@ package com.microsoft.durabletask.exporthistory;
 import com.microsoft.durabletask.DurableTaskClient;
 import com.microsoft.durabletask.ListInstanceIdsQuery;
 import com.microsoft.durabletask.ListInstanceIdsResult;
+import com.microsoft.durabletask.OrchestrationRuntimeStatus;
 import com.microsoft.durabletask.TaskActivity;
 import com.microsoft.durabletask.TaskActivityContext;
 
+import javax.annotation.Nullable;
+import java.time.Instant;
 import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Activity that lists terminal orchestration instances for a completion-time window using the client
  * {@code listInstanceIds} wrapper, returning a page plus the checkpoint to advance to.
  */
-public final class ListTerminalInstancesActivity implements TaskActivity {
+final class ListTerminalInstancesActivity implements TaskActivity {
 
     /** The registered activity name. */
     public static final String NAME = "ListTerminalInstancesActivity";
@@ -45,5 +49,171 @@ public final class ListTerminalInstancesActivity implements TaskActivity {
         return new InstancePage(
                 new ArrayList<>(result.getInstanceIds()),
                 new ExportCheckpoint(result.getContinuationToken()));
+    }
+}
+
+/**
+ * Input to {@link ListTerminalInstancesActivity}: a completion-time window, terminal status filter, pagination
+ * cursor, and batch size.
+ */
+final class ListTerminalInstancesRequest {
+
+    private Instant completedTimeFrom;
+    private Instant completedTimeTo;
+    private List<OrchestrationRuntimeStatus> runtimeStatus;
+    private String lastInstanceKey;
+    private int maxInstancesPerBatch;
+
+    /** Creates an empty {@code ListTerminalInstancesRequest} (for deserialization). */
+    public ListTerminalInstancesRequest() {
+    }
+
+    /**
+     * Creates a {@code ListTerminalInstancesRequest}.
+     *
+     * @param completedTimeFrom    the inclusive completion-time lower bound
+     * @param completedTimeTo      the inclusive completion-time upper bound, or {@code null}
+     * @param runtimeStatus        the terminal runtime statuses, or {@code null}
+     * @param lastInstanceKey      the pagination cursor from the previous page, or {@code null}
+     * @param maxInstancesPerBatch the maximum number of instance IDs per page
+     */
+    public ListTerminalInstancesRequest(
+            Instant completedTimeFrom,
+            @Nullable Instant completedTimeTo,
+            @Nullable List<OrchestrationRuntimeStatus> runtimeStatus,
+            @Nullable String lastInstanceKey,
+            int maxInstancesPerBatch) {
+        this.completedTimeFrom = completedTimeFrom;
+        this.completedTimeTo = completedTimeTo;
+        this.runtimeStatus = runtimeStatus;
+        this.lastInstanceKey = lastInstanceKey;
+        this.maxInstancesPerBatch = maxInstancesPerBatch;
+    }
+
+    /** @return the inclusive completion-time lower bound. */
+    public Instant getCompletedTimeFrom() {
+        return this.completedTimeFrom;
+    }
+
+    /**
+     * Sets the inclusive completion-time lower bound.
+     *
+     * @param completedTimeFrom the lower bound
+     */
+    public void setCompletedTimeFrom(Instant completedTimeFrom) {
+        this.completedTimeFrom = completedTimeFrom;
+    }
+
+    /** @return the inclusive completion-time upper bound, or {@code null}. */
+    @Nullable
+    public Instant getCompletedTimeTo() {
+        return this.completedTimeTo;
+    }
+
+    /**
+     * Sets the inclusive completion-time upper bound.
+     *
+     * @param completedTimeTo the upper bound, or {@code null}
+     */
+    public void setCompletedTimeTo(@Nullable Instant completedTimeTo) {
+        this.completedTimeTo = completedTimeTo;
+    }
+
+    /** @return the terminal runtime statuses, or {@code null}. */
+    @Nullable
+    public List<OrchestrationRuntimeStatus> getRuntimeStatus() {
+        return this.runtimeStatus;
+    }
+
+    /**
+     * Sets the terminal runtime statuses.
+     *
+     * @param runtimeStatus the runtime statuses, or {@code null}
+     */
+    public void setRuntimeStatus(@Nullable List<OrchestrationRuntimeStatus> runtimeStatus) {
+        this.runtimeStatus = runtimeStatus;
+    }
+
+    /** @return the pagination cursor from the previous page, or {@code null}. */
+    @Nullable
+    public String getLastInstanceKey() {
+        return this.lastInstanceKey;
+    }
+
+    /**
+     * Sets the pagination cursor.
+     *
+     * @param lastInstanceKey the cursor, or {@code null}
+     */
+    public void setLastInstanceKey(@Nullable String lastInstanceKey) {
+        this.lastInstanceKey = lastInstanceKey;
+    }
+
+    /** @return the maximum number of instance IDs per page. */
+    public int getMaxInstancesPerBatch() {
+        return this.maxInstancesPerBatch;
+    }
+
+    /**
+     * Sets the maximum number of instance IDs per page.
+     *
+     * @param maxInstancesPerBatch the batch size
+     */
+    public void setMaxInstancesPerBatch(int maxInstancesPerBatch) {
+        this.maxInstancesPerBatch = maxInstancesPerBatch;
+    }
+}
+
+/**
+ * Output of {@link ListTerminalInstancesActivity}: a page of terminal instance IDs plus the checkpoint to advance
+ * to once this page has been exported.
+ */
+final class InstancePage {
+
+    private List<String> instanceIds = new ArrayList<>();
+    private ExportCheckpoint nextCheckpoint;
+
+    /** Creates an empty {@code InstancePage} (for deserialization). */
+    public InstancePage() {
+    }
+
+    /**
+     * Creates an {@code InstancePage}.
+     *
+     * @param instanceIds    the page of terminal instance IDs
+     * @param nextCheckpoint the checkpoint to advance to after exporting this page, or {@code null}
+     */
+    public InstancePage(List<String> instanceIds, @Nullable ExportCheckpoint nextCheckpoint) {
+        this.instanceIds = instanceIds;
+        this.nextCheckpoint = nextCheckpoint;
+    }
+
+    /** @return the page of terminal instance IDs. */
+    public List<String> getInstanceIds() {
+        return this.instanceIds;
+    }
+
+    /**
+     * Sets the page of terminal instance IDs.
+     *
+     * @param instanceIds the instance IDs
+     */
+    public void setInstanceIds(List<String> instanceIds) {
+        this.instanceIds = instanceIds;
+    }
+
+    /** @return the checkpoint to advance to after exporting this page, or {@code null}. */
+    @Nullable
+    public ExportCheckpoint getNextCheckpoint() {
+        return this.nextCheckpoint;
+    }
+
+    /**
+     * Sets the checkpoint to advance to after exporting this page.
+     *
+     * @param nextCheckpoint the next checkpoint, or {@code null}
+     */
+    public void setNextCheckpoint(@Nullable ExportCheckpoint nextCheckpoint) {
+        this.nextCheckpoint = nextCheckpoint;
     }
 }
